Moved 2d map validation to correct file
